Fixed Topology Resolution Issue
