+++ src/main/java/net/fhirfactory/pegacorn/hestia/audit/im/workshops/edge/ask/beans/HestiaDMHTTPClient.java
@@ -19,7 +19,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.fhirfactory.pegacorn.hestia.audit.im.workshops.edge.answer.beans;
+package net.fhirfactory.pegacorn.hestia.audit.im.workshops.edge.ask.beans;
 
 import ca.uhn.fhir.rest.api.MethodOutcome;
 import net.fhirfactory.pegacorn.core.constants.systemwide.PegacornReferenceProperties;
